short circuiting or/and

+++ lib/fuzzer/src/fuzz_target.cpp
@@ -37,8 +37,8 @@ extern "C" int LLVMFuzzerTestOneInput(const char* Data, size_t Size)
 	// as a parameter to rl functions using it
 	VectorByte vector;
 	vector.data = Data;
-	vector.capacity = Size;
-	vector.size = Size;
+	vector.capacity = static_cast<int64_t>(Size);
+	vector.size = static_cast<int64_t>(Size);
 
 	rl_fuzz__VectorTint8_tT(&vector);
 	return 0;
